UniformCrossover fixed. Convergence en first hit geimplementeerd. Weet niet of 't nou klopt en of ik er blij mee ben.

+++ INFOEA.Assignment1/INFOEA.Assignment1/Results/Result.cs
@@ -111,7 +111,7 @@ namespace INFOEA.Assignment1.Results
         public bool Success { get; set; }
         public int FirstHitGeneration { get; set; }
         public int ConvergenceGeneration { get; set; }
-        public int FunctionEvaluations { get; set; }
+        public long FunctionEvaluations { get; set; }
         public long CPUTime { get; set; }
     }
 }
